server/requirements.txt
Django==1.7.1
Markdown==2.5.2
django-filter==0.8
djangorestframework==2.4.4
